Enhanced Modifiers
Resolved the underlying design issue which used stateful lambdas to
capture the modifiers for comparison. The new design uses an alternate
model which separates the declaration of modifiers from the matcher.
Instead of `Signature.builder().modifiers(any(...))`, it is changed to
`Signature.builder().modifiers(...).contains(ANY)`. In my opinion, it is
a good compromise between API design and performance.