Rough prototype of annotations framework
Created a rough prototype of the annotations framework which lacks
support for annoations at the moment. Whether they will be supported is
a completely separate issue which needs to be tackled. At present, the
signature for variables is completed although I'm concerned with the
memory overhead incurred by the use of lambdas. Also, I forgot to commit
yesterday. Opps.