Allowed methods to be inferred as CommandExecutors.
There are instances in which it is beneficial for several CommandExecutors to share state. As such, methods can now be infered as CommandExecutors so long as
they match the CommandExecutor interface and consequently, a class to contain several CommandExecutors. Additionally the class which contains the methods must
implement the CommandExecutors marker interface and that the command annotations fully supports the new addition. Implementation-wise, a method reference is generated from
each method which is annotated with the @Namespace annotation and is after registration, treated equally as a normal CommandExecutor. Consequently the changes made to support the
new feature has been minimal and only requires the annotations to also target methods and for the CommandResolver to expose a new method for registering CommandExecutors instances.
The only singificant change is the Resolver interface which now requires an additional AnnotatedElement. There is minimal performance penalty thus far for due to the implementation
generating a method refernce instead of constatly calling the method via reflection.