Refactored signature package to implement the TreeVisitor interface
Refactored the types to use the TreeVisitor interface which removes the
need to manually check and cast the trees to a given type. Additionally
provides the benefit of condensing the required classes and boilerplate
code while providing an overall better memory footprint due to the
TreeVisitors being stateless for the most parts. The only downside is
that the current design requires the class to be passed to the
type/visitor itself but is an acceptable tradeoff.