Migrated project to Java 9.
Since migration to Java 9 has been completed, the @JDK9 annotation has been removed. In actual reality, the syntax sugar added in Java 9 was underwhelming to say the least. Only a small fraction of the codebase could be migrated to use the new styled try-with statements. It was extremely misleading in the JEP as it was assumed that it also accepeted expressions which returned effectively final values. The diamond inference could not be widely applied due to the classes being wrapped in spy() methods which disallowed the typeparameters from being infered. Additionally, the codebase is currently under review for simiplification opportunities which resulted in the CommandExecutors interface being removed.