More progress on Locales package
Finished most of the initial prototype. I realised how god awful the
previous Locales implementation was. It was extremely inefficent and
worse still, wrong. Hopefully, I did a better job this time around.

Configuration needs to be updated to support writing to files. Thus, it
consists of 2 portions which can be separated. Mapping and
serialization.
The current implementation needs to be split into ConfiguationMapper and
Configuration.

TODO
- Add PlayerConfigurationProvider
- Split Configuration into Configuration and ConfigurationMapper
- Move configuration to separate package
- Merge PlayerResource with Resource?