Reduce scope of Chimera
Limited resources and time will mean that the scope of Chimera has to be
reduced for even a slight chance of a full release. Simply put, we don't
have enough manpower to maintain some of the packages in the library,
hence the removal.

The packages to be removed are the util.codec.*, util.concurrent.bukkit
packages. In addition, the inclusion of a GUI package, annotations &
maven-plugin has been put on hold. Focus will be directed towards the
command, particles & refactoring the existing util.conucrrent package.